CA-193907: refresh consoles when host address set
During startup, the host address in the database is set before
Xapi_mgmt_iface.on_dom0_networking_change is called, and therefore it
does not detect the address change, and does not refresh the console
URLs.

This commit tries to keep the host address and its console URL in the
database in sync, by adding a call to
Dbsync_master.refresh_console_urls, when a host's IP address is updated
in the database, like it's already done in Xapi_pool.hello (to ensure
that the dom0 console IPs are consistent with the host addresses in the
database).

The functions had to be reordered in Dbsync_master to make it compile.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>