CA-242706: run update-issue only at mgmt if change
Only run the update-issue script when the management interface changes.
This accompanies another commit, which changes the update-issue script
so that /etc/issue will automatically print out the IP address for the
specified interface. Therefore we only need to run it when the
management interface changes, but not every time when the management IP
changes.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>