CA-242706: remove calls to update-getty
This accompanies another commit, which changes update-issue to generate
an /etc/issue, which automatically determines the IP address of the
machine using the \4 agetty escape code.

A possible problem with this is that it may not print out the IP address
for the management interface, and I'm not sure what it will print out
when no IP address is configured.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>