Test_vdi_cbt: test that data_destroy waits for VBDs to be unplugged
The usual workflow for backup vendors is that they connect to the VDI
snapshot via NBD, download the changed blocks, then disconnect, and run
data_destroy on the VDI.

However, the NBD client and server do not wait until the cleanup actions
finish, they immediately disconnect. Therefore running a
VDI.data_destroy immediately after disconnecting results in a VDI_IN_USE
error, because it takes time for the VBD.unplug and destroy calls to
complete.

To improve the usability of the CBT XenAPI calls, we'll make
data_destroy wait until the VBDs get unplugged, in case all the attached
VBDs of the VDI have an unplug operation in progress.

Signed-off-by: Iglói Gábor <gaborigloi@gmail.com>