Update VDI.clone,snapshot SMAPIv2 calls to return correct cbt_enabled value
Now that cbt_enabled is part of the vdi_info type at the SMAPIv2 layers,
we have to make sure that it is correctly set by these functions at the
SMAPIv2 layer too, as described by the design. This is to ensure that
clients using SMAPIv2 functions directly will get back correct vdi_info
values.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>