Update VDI.enable_cbt,disable_cbt SMAPIv2 calls to set cbt_enabled value
Now that cbt_enabled is part of the vdi_info type at the SMAPIv2 layers,
we have to make sure that it is correctly set by these functions at the
SMAPIv2 layer too, as described by the design. This is to ensure that
SMAPIv2 clients using these SMAPIv2 functions directly will get back
correct vdi_info values if they inspect the VDI after running one of
these operations.

We have to be careful when manipulating xapi's database in the SMAPIv2
layer, because this layer is backed by xapi's DB, and the XenAPI layer
also changes the DB and treats SMAPIv2 as an abstract interface, and
this could lead to conflicts. These DB accesses I've added do not
conflict with what the corresponding XenAPI functions do, the VDI's
cbt_enabled field will just get updated a bit earlier by the SMAPIv2
layer too. I decided to still set this field in the XenAPI functions
too, because there we treat SMAPIv2 as an abstract interface and
therefore do not know that it actually manipulates xapi's DB under the
hood, so we have to set the cbt_enabled field to make sure it has the
correct value.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>