Update allowed operations when creating Cluster & Cluster_host
I've added the calls to update_allowed_operations to the message
forwarding layer: Cluster_host.create is redirected to the host itself,
and therefore if we did this call in the implementation in
xapi_cluster_host.ml, that would result in DB calls to the master.

These update_allowed_operations should be safe and should not throw an
exception, because that would cause all their callers to fail.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>