Add prechecks to shutdown, reboot XenAPI calls to fail early
This helps us preserve the original behaviour.

We should decide whether we to keep the assertion that the host is disabled in
prepare_for_shutdown, because in case the user just runs "reboot",
nothing will be done due to the failure of this initial assertion, and
the host will continue to shut down.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>