CP-29757: block SXM of encrypted VDIs
This will not block the "Move VM" operation on a halted VM in XenCenter,
because XenCenter moves the VM manually by copying the VDIs of the VM
and creating & plugging new VBDs.

But this is fine, because we wanted to block SXM of encrypted VDIs for
XenAPI users. And both the VM.migrate_send call and "xe vm-migrate" will
be blocked for encrypted VDIs, even if the VM is halted.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>
(cherry picked from commit 37519e16c84878db6184c4525219293cfe6ab06e)

For this cherry-pick I had to include some earlier cleanup commits.